server/requirements.txt
flask==1.1.2
flask_accept==0.0.6
paho-mqtt==1.5.1
psycopg2==2.8.6
